requirements.txt
pydantic>=2.0.0
langgraph>=0.0.15
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.0.0
graphviz>=0.20.1
python-dotenv>=1.0.0 
python-dotenv>=1.0.0